Safari-611.1.10/LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-zh-Hans-expected.txt

PASS navigator.language is "zh-cn"
PASS successfullyParsed is true

TEST COMPLETE
